+++ ApiBiblioteca.Application/Interfaces/IRepository/ICategoriaRepository.cs
@@ -6,8 +6,8 @@ namespace ApiBiblioteca.Application.Interfaces.IRepository;
 public interface ICategoriaRepository
 {
     Task<PagedList<Categoria>> GetAllAsync(int skip, int take);
-    Task<PagedList<Livro>> GetLivrosByCategoriaAsync(long categoriaId, int skip, int take);
-    Task<Categoria> GetByIdAsync(long categoriaId);
+    Task<PagedList<Livro>> GetLivrosByCategoriaAsync(int categoriaId, int skip, int take);
+    Task<Categoria> GetByIdAsync(int categoriaId);
     void Create(Categoria categoria);
     void Remove(Categoria categoria);
 }
