+++ ApiBiblioteca.Application/Interfaces/IRepository/ILivroRepository.cs
@@ -6,8 +6,8 @@ namespace ApiBiblioteca.Application.Interfaces.IRepository;
 public interface ILivroRepository
 {
     Task<PagedList<Livro>> GetAllAsync(int skip, int take);
-    Task<PagedList<ExemplarLivro>> GetExemplaresByLivroAsync(long livroId, int skip, int take);
-    Task<Livro> GetByIdAsync(long id);
+    Task<PagedList<ExemplarLivro>> GetExemplaresByLivroAsync(int livroId, int skip, int take);
+    Task<Livro> GetByIdAsync(int id);
     Task<bool> ExistsByIsbn(string isbn);
     void Create(Livro livro);
     void Remove(Livro livro);
