+++ ApiBiblioteca.Application/Interfaces/IServices/IAutorService.cs
@@ -8,9 +8,9 @@ namespace ApiBiblioteca.Application.Interfaces.IServices;
 public interface IAutorService
 {
     public Task<PagedList<AutorResponseDto>> Get(QueryParameters parameters);
-    public Task<PagedList<LivroResponseDto>> GetComLivros(long autorId, QueryParameters parameters);
-    public Task<AutorResponseDto> GetId(long id);
+    public Task<PagedList<LivroResponseDto>> GetComLivros(int autorId, QueryParameters parameters);
+    public Task<AutorResponseDto> GetId(int id);
     public Task<AutorResponseDto> Create(AutorDto dto);
-    public Task<AutorResponseDto> Update(long id, AutorDto dto);
-    public Task Delete(long id);
+    public Task<AutorResponseDto> Update(int id, AutorDto dto);
+    public Task Delete(int id);
 }
