+++ ApiBiblioteca.Application/Interfaces/IServices/ILivroService.cs
@@ -8,9 +8,9 @@ namespace ApiBiblioteca.Application.Interfaces.IServices;
 public interface ILivroService
 {
     public Task<PagedList<LivroResponseDto>> Get(QueryParameters parameters);
-    public Task<PagedList<ExemplarResponseDto>> GetComExemplares(long livroId, QueryParameters parameters);
-    public Task<LivroResponseDto> GetId(long livroId);
+    public Task<PagedList<ExemplarResponseDto>> GetComExemplares(int livroId, QueryParameters parameters);
+    public Task<LivroResponseDto> GetId(int livroId);
     public Task<LivroResponseDto> Create(CreateLivroDto dto);
-    public Task<LivroResponseDto> Update(long livroId, UpdateLivroDto dto);
-    public Task Delete(long livroId);
+    public Task<LivroResponseDto> Update(int livroId, UpdateLivroDto dto);
+    public Task Delete(int livroId);
 }
